Bug fixing day 2

+++ uploadServer.py
@@ -7,6 +7,7 @@ import time as t
 import os
 import inspect
 from scripts import makeFile
+
 cwd = os.getcwd()
 UPLOAD_FOLDER = cwd + '\\datasets'
 external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
@@ -14,27 +15,32 @@ app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
 app.layout = html.Div(
-        children=[
-            html.Iframe(id='iframe-upload',src=f'/upload'),
-            html.Div(id='output')
-                ]
+    children=[
+        html.Iframe(id='iframe-upload', src=f'/upload'),
+        html.Div(id='output')
+    ]
 )
 
 
 @app.server.route('/upload', methods=['GET', 'POST'])
 def upload_file():
     if request.method == 'POST':
-        now = t.time()
-        file = request.files['file']
-        filename = secure_filename(file.filename)
-        file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-        if ".csv.gz" in filename:
-            makeFile.csv_gz_to_txt(filename)
-            os.remove(UPLOAD_FOLDER+"/"+filename)
-        if ".dat_.gz" in filename:
-            makeFile.dat_gz_to_txt(filename)
-            os.remove(UPLOAD_FOLDER+"/"+filename)
-        print(t.time(), "@", inspect.currentframe().f_code.co_name, "<<<UPLOADING TOOK",t.time()-now,"SECONDS>>>")
+        try:
+            now = t.time()
+            file = request.files['file']
+            filename = secure_filename(file.filename)
+            if ".csv.gz" in filename or ".dat_.gz" in filename or '.txt' in filename:
+                file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+                if ".csv.gz" in filename:
+                    makeFile.csv_gz_to_txt(filename)
+                    os.remove(UPLOAD_FOLDER + "/" + filename)
+                elif ".dat_.gz" in filename:
+                    makeFile.dat_gz_to_txt(filename)
+                    os.remove(UPLOAD_FOLDER + "/" + filename)
+            print(t.time(), "@", inspect.currentframe().f_code.co_name, "<<<UPLOADING TOOK", t.time() - now,
+                  "SECONDS>>>")
+        except KeyError:
+            pass
     return '''
     <form method=post enctype=multipart/form-data>
       <input type=file name=file>
@@ -44,4 +50,4 @@ def upload_file():
 
 
 if __name__ == '__main__':
-   app.run_server(debug=True, port=8051)
+    app.run_server(debug=True, port=8051)
